# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger MusicalShop Documentation",
        "version": "1.0.0"
    },
    "paths": {
...
        "/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Регистрация нового пользователя",
                "operationId": "registerUser",
                "requestBody": {
                    "description": "Данные для регистрации",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Ruslan"
                                    },
                                    "surname": {
                                        "type": "string",
                                        "example": "Farvaev"
                                    },
                                    "patronymic": {
                                        "type": "string",
                                        "example": "Rinatovich"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "+79990011222"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "Ruslan@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password"
                                    },
                                    "Confirm password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Успешная регистрация"
                    },
                    "400": {
                        "description": "Ошибка валидации"
                    }
                }
            }
        },
        "/basket": {
            "get": {
                "tags": [
                    "Basket"
                ],
                "summary": "Получить содержимое корзины",
                "operationId": "basket",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/basket/confirm": {
            "post": {
                "tags": [
                    "Basket"
                ],
                "summary": "Подтверждение корзины",
                "operationId": "basketConfirm",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "123456789"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Validation Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/basket/add/{product_id}": {
            "post": {
                "tags": [
                    "Basket"
                ],
                "summary": "Добавление товара в корзину",
                "operationId": "basketAdd",
                "parameters": [
                    {
                        "name": "product_id",
                        "in": "path",
                        "description": "ID товара",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/basket/remove/{product_id}": {
            "delete": {
                "tags": [
                    "Basket"
                ],
                "summary": "Удаление товара из корзины",
                "operationId": "basketRemove",
                "parameters": [
                    {
                        "name": "product_id",
                        "in": "path",
                        "description": "ID товара",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}